add styles to the phone component/fix bugs
