+++ src/app/inner-pages/dev-pages/dev-profile/dev-contact-info/dev-contact-info-location-edit/dev-contact-info-location-edit.component.html
@@ -28,14 +28,9 @@
     <i class="phone phone-gray"></i>
     <span class="text text__sm text__bold text__uppercase">PHONE</span>
 
-<!--   <app-input [type]="'inner'"-->
-<!--               formControlName="phone"-->
-<!--               [placeholder]="'Phone'"-->
-<!--               [value]="form.get('phone').value"></app-input>-->
-    <app-input-test [type]="'inner'"
-               formControlName="phone"
+    <app-phone formControlName="phone"
                [placeholder]="'Phone'"
-               [value]="form.get('phone').value"></app-input-test>
+               [value]="form.get('phone').value"></app-phone>
 
   </div>
 </div>
