--- src/app/shared/components/input-test/input-test.component.html
@@ -1,30 +1,0 @@
-<!--<input-->
-<!--  (keyup.enter)="enterKeyClick()"-->
-<!--  (focus)="onFocus()"-->
-<!--  class="{{ type }}"-->
-<!--  [type]="isHiddenPassword ? 'password' : 'text'"-->
-<!--  [(ngModel)]="value"-->
-<!--  [placeholder]="placeholder"-->
-<!--&gt;-->
-
-<ngx-intl-tel-input
-  [cssClass]="'custom'"
-  [preferredCountries]="preferredCountries"
-  [enableAutoCountrySelect]="true"
-  [enablePlaceholder]="true"
-  [searchCountryFlag]="true"
-  [searchCountryField]="[SearchCountryField.Iso2, SearchCountryField.Name]"
-  [selectFirstCountry]="false"
-  [selectedCountryISO]="CountryISO.UnitedKingdom"
-  [maxLength]="15"
-  [tooltipField]="TooltipLabel.Name"
-  [phoneValidation]="true"
-  [separateDialCode]="separateDialCode"
-  [numberFormat]="PhoneNumberFormat.National"
-  [(ngModel)]="value"
-  name="phone">
-</ngx-intl-tel-input>
-
-
-
-<ng-content></ng-content>
