+++ src/app/shared/components/phone/phone.component.scss
@@ -1,99 +1,97 @@
-// /* You can add global styles to this file, and also import other style files */
-// body {
-// 	font-family: Roboto, Helvetica, Arial, sans-serif;
-// }
+@import 'assets/scss/variables';
 
-// .wrapper {
-// 	margin-bottom: 20px;
-// }
+::ng-deep {
+  .iti {
+    display: inline-block;
+    width: 232px;
+    height: 40px;
+    border: 1px solid $color-4;
+    border-radius: 5px;
+    background: transparent;
+    font-family: Raleway;
+    font-feature-settings: "lnum";
+    font-size: 12px;
+  }
 
-// /*****************************************************
-// 	Custom styling example bellow.
-// *****************************************************/
+  .iti .dropdown-menu.country-dropdown {
+    border-top-left-radius: 0px;
+    border-top-right-radius: 0px;
+    border-color: #c7cace;
+    margin-top: -1px;
+  }
 
-// .iti {
-// 	display: block !important;
-// 	margin-bottom: 20px;
-// }
+  .iti .iti__country-list {
+    box-shadow: none;
+    font-size: 14px;
+    margin-left: 0;
+    width: 244px;
+    max-height: 170px;
+  }
 
-// .iti .dropdown-menu.country-dropdown {
-// 	border-top-left-radius: 0px;
-// 	border-top-right-radius: 0px;
-// 	border-color: #c7cace;
-// 	margin-top: -1px;
-// }
+  .iti__flag-container.open + input {
+    border-bottom-left-radius: 0px;
+    border-bottom-right-radius: 0px;
+  }
 
-// .iti .iti__country-list {
-// 	box-shadow: none;
-// 	font-size: 14px;
-// 	margin-left: 0;
-// 	width: 244px;
-// 	max-height: 170px;
-// }
+  .iti .search-container input {
+    font-size: 14px;
+    border-color: #c7cace;
+    border-radius: 0;
+    padding: 5px 10px;
+  }
 
-// .iti__flag-container.open + input {
-// 	border-bottom-left-radius: 0px;
-// 	border-bottom-right-radius: 0px;
-// }
+  .iti .search-container input:focus {
+    outline: none;
+  }
 
-// .iti .search-container input {
-// 	font-size: 14px;
-// 	border-color: #c7cace;
-// 	border-radius: 0;
-// 	padding: 5px 10px;
-// }
+  @media screen and (max-width: 479px) {
+    .iti .iti__country-list {
+      width: 88.3vw;
+    }
+  }
 
-// .iti .search-container input:focus {
-// 	outline: none;
-// }
+  ngx-intl-tel-input input {
+    height: 40px;
+    padding: 10px;
+    border-style: solid;
+    border-color: #c7cace;
+    border-radius: 5px;
+    font-size: 18px;
+    box-shadow: 0 4px 4px rgba(0, 0, 0, 0.5)
+  }
 
-// @media screen and (max-width: 479px) {
-// 	.iti .iti__country-list {
-// 		width: 88.3vw;
-// 	}
-// }
+  ngx-intl-tel-input.ng-invalid.ng-touched input {
+    border: 2px solid #c0392b;
+  }
 
-// ngx-intl-tel-input input {
-// 	height: 44px;
-// 	margin-bottom: 20px;
-// 	padding: 10px;
-// 	border-style: solid;
-// 	border-width: 1px;
-// 	border-color: #c7cace;
-// 	border-radius: 4px;
-// 	font-size: 18px;
-// }
+  ngx-intl-tel-input input:hover {
+  }
 
-// ngx-intl-tel-input.ng-invalid.ng-touched input {
-// 	border: 1px solid #c0392b;
-// }
+  ngx-intl-tel-input input:focus {
+    outline: none !important;
+    border-color: #3498db;
+    box-shadow: 0 0 0 0 #000;
+  }
 
-// ngx-intl-tel-input input:hover {
-// 	box-shadow: 0 1px 0 0 rgba(0, 0, 0, 0.24);
-// }
+  ngx-intl-tel-input input::-webkit-input-placeholder {
+    color: #bac2c7;
+  }
 
-// ngx-intl-tel-input input:focus {
-// 	outline: none !important;
-// 	border-color: #3498db;
-// 	box-shadow: 0 0 0 0 #000;
-// }
+  ngx-intl-tel-input input:-ms-input-placeholder {
+    color: #bac2c7;
+  }
 
-// ngx-intl-tel-input input::-webkit-input-placeholder {
-// 	color: #bac2c7;
-// }
+  ngx-intl-tel-input input::-ms-input-placeholder {
+    color: #bac2c7;
+  }
 
-// ngx-intl-tel-input input:-ms-input-placeholder {
-// 	color: #bac2c7;
-// }
+  ngx-intl-tel-input input::placeholder {
+    color: #bac2c7;
+  }
 
-// ngx-intl-tel-input input::-ms-input-placeholder {
-// 	color: #bac2c7;
-// }
+  ngx-intl-tel-input input[disabled] {
+    background-color: #e5eaf1;
+  }
+}
 
-// ngx-intl-tel-input input::placeholder {
-// 	color: #bac2c7;
-// }
 
-// ngx-intl-tel-input input[disabled] {
-// 	background-color: #e5eaf1;
-// }
